Provisional support for COPAC, and other servers using Dublin Core XML.  Rudimentary parser (currently private to the zoom group server) uses NSXMLDocument, and is selected based in serverInfo options.
Fix copy method on serverInfo, and add an options dictionary.  We may
want to switch options over to ivars, if we determine the necessary
members.  Currently the serverInfo's options ivar is only created by the plist,
since there's no UI for those.

Changed binding of sheet options directly to serverInfo ivar, and add
a convenience method for creating a new serverInfo.  This removes the
duplicate ivars (which hopefully is okay), and all operations are bound
to a copy or a newly created serverInfo.  UI still seems to work OK.
This moves validation into the serverInfo class.  Minimally tested with
new and old z39.50 groups, seems to work.


git-svn-id: http://svn.code.sf.net/p/skim-app/code/trunk@269 0fceea05-a30d-0410-8a8b-80ef821fb0a1