Use PDFAnnotations for our notes. Save them as dictionaries in the xattrs. For now we can read old-style notes from xattrs, I think.
Use a custom subclass of PDFAnnotationText for notes displayed as icons, supporting attributed text and images. 

Add some support for other annotation types, not accessible from the UI at the moment (I am not sure if we should). 

Rework note window to support annotations and run as a panel. Shown on a double-click of a Note annotation. Add BDSKDragImagView copied from bibdesk. 

Change linewidth of circle annotations, and some colors. 

Additional support for reworking and improvements. 

git-svn-id: http://svn.code.sf.net/p/skim-app/code/skim@868 0fceea05-a30d-0410-8a8b-80ef821fb0a1